petits tests drivers et corrections mineures

+++ Peripheriques/Accelerometre.c
@@ -1,6 +1,7 @@
 #include "Accelerometre.h"
 #include <math.h>
 
+
 void GPIOConfAccelero(){
 	MyGPIO_pin_conf(GPIOC, 0, 'a');
 	MyGPIO_pin_conf(GPIOC, 1, 'a');
@@ -11,7 +12,7 @@ void ADCConfAccelero(){
 }
 
 int getAngleRoulis(){
-	int valueADC, angleRoulis;
+	double valueADC, angleRoulis;
 	valueADC = MyADC_get_value(ADC1);
 	angleRoulis = acos(valueADC/0.48);
 	return angleRoulis;
